Math Changes.

+++ src/main/java/com/worldcretornica/plotme/defaultgenerator/DefaultRoadPopulator.java
@@ -42,14 +42,14 @@ public class DefaultRoadPopulator extends BlockPopulator {
         int mod1 = 1;
 
         if (pathsize % 2 == 1) {
-            n1 = Math.ceil(((double) pathsize) / 2) - 2;
-            n2 = Math.ceil(((double) pathsize) / 2) - 1;
-            n3 = Math.ceil(((double) pathsize) / 2);
+            n1 = Math.ceil(pathsize / 2) - 2;
+            n2 = Math.ceil(pathsize / 2) - 1;
+            n3 = Math.ceil(pathsize / 2);
             mod2 = -1;
         } else {
-            n1 = Math.floor(((double) pathsize) / 2) - 2;
-            n2 = Math.floor(((double) pathsize) / 2) - 1;
-            n3 = Math.floor(((double) pathsize) / 2);
+            n1 = Math.floor(pathsize / 2) - 2;
+            n2 = Math.floor(pathsize / 2) - 1;
+            n3 = Math.floor(pathsize / 2);
         }
 
         for (int x = xx; x < xx + 16; x++) {
